- Adopt flagged_spatial_id (full implementation), cat_model parameter,
  and schema-based column descriptions from main
- Adopt open_dataset, inject_seed, compute_var_rnoise, float32 err cast
- Thread cat_model through catalog_generator.py and process_detection_image

+++ romancal/source_catalog/utils.py
@@ -84,7 +84,8 @@ def copy_model_arrays(model):
         copied_model = ImageModel()
         copied_model.meta = model.meta
         copied_model.data = model.data.copy()
-        copied_model.err = model.err.copy()
+        # cast to float32 so unit manipulations later on don't overflow
+        copied_model.err = model.err.copy().astype("float32")
         copied_model.dq = model.dq
     elif isinstance(model, MosaicModel):
         copied_model = MosaicModel()
